requirements.txt
# Core dependencies for Bharat Heritage Virtual Try-On
gradio>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# Google Cloud AI Platform (optional - for full AI functionality)
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0
google-cloud-logging>=3.8.0

# Optional dependencies for enhanced functionality
numpy>=1.24.0
opencv-python>=4.8.0